Use CDT packages during testing (for X11)
Doing it this way means that we do not need X11 installed on the test machine,
but instead requires that we use LD_LIBRARY_PATH. An alternative may be to use
an overlay/union filesystem to merge the sysroot with /usr/lib (or /usr/lib64)

If we used an overlay filesystem we could run the tests twice, with inverted
priority on the unioning. This would check compatibility with both the CDT and
the native sysroot.